+++ integration-specs/simple/spec/spec_helper.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'bundler/setup'
+Bundler.setup
+
+ENV['SEQUENT_ENV'] ||= 'test'
+
+require_relative '../simple'