+++ lib/sequent/configuration.rb
@@ -36,6 +36,7 @@ module Sequent
 
     attr_accessor :aggregate_repository,
                   :event_store,
+                  :event_store_cache_event_types,
                   :command_service,
                   :event_record_class,
                   :stream_record_class,
@@ -63,7 +64,8 @@ module Sequent
                   :enable_multiple_database_support,
                   :primary_database_role,
                   :primary_database_key,
-                  :time_precision
+                  :time_precision,
+                  :enable_autoregistration
 
     attr_reader :migrations_class_name,
                 :versions_table_name,
@@ -73,10 +75,13 @@ module Sequent
       @instance ||= new
     end
 
+    # Create a new instance of Configuration
     def self.reset
       @instance = new
     end
 
+    # Restore the given Configuration
+    # @param configuration [Sequent::Configuration]
     def self.restore(configuration)
       @instance = configuration
     end
@@ -88,6 +93,7 @@ module Sequent
       self.command_middleware = Sequent::Core::Middleware::Chain.new
 
       self.aggregate_repository = Sequent::Core::AggregateRepository.new
+      self.event_store_cache_event_types = true
       self.event_store = Sequent::Core::EventStore.new
       self.command_service = Sequent::Core::CommandService.new
       self.event_record_class = Sequent::Core::EventRecord
@@ -121,6 +127,8 @@ module Sequent
       self.primary_database_key = :primary
 
       self.time_precision = DEFAULT_TIME_PRECISION
+
+      self.enable_autoregistration = false
     end
 
     def can_use_multiple_databases?
@@ -148,6 +156,54 @@ module Sequent
       end
 
       @migrations_class_name = class_name
+    end
+
+    def autoregister!
+      return unless enable_autoregistration
+
+      # Only autoregister the AggregateSnapshotter if the autoregistration is enabled
+      Sequent::Core::AggregateSnapshotter.skip_autoregister = false
+
+      Rails.autoloaders.main.eager_load(force: true) if defined?(Rails)
+
+      self.class.instance.command_handlers ||= []
+      for_each_autoregisterable_descenant_of(Sequent::CommandHandler) do |command_handler_class|
+        Sequent.logger.debug("[Configuration] Autoregistering CommandHandler #{command_handler_class}")
+        self.class.instance.command_handlers << command_handler_class.new
+      end
+
+      self.class.instance.event_handlers ||= []
+      for_each_autoregisterable_descenant_of(Sequent::Projector) do |projector_class|
+        Sequent.logger.debug("[Configuration] Autoregistering Projector #{projector_class}")
+        self.class.instance.event_handlers << projector_class.new
+      end
+
+      for_each_autoregisterable_descenant_of(Sequent::Workflow) do |workflow_class|
+        Sequent.logger.debug("[Configuration] Autoregistering Workflow #{workflow_class}")
+        self.class.instance.event_handlers << workflow_class.new
+      end
+
+      self.class.instance.command_handlers.map(&:class).tally.each do |(clazz, count)|
+        if count > 1
+          fail "CommandHandler #{clazz} is registered #{count} times. A CommandHandler can only be registered once"
+        end
+      end
+
+      self.class.instance.event_handlers.map(&:class).tally.each do |(clazz, count)|
+        if count > 1
+          fail "EventHandler #{clazz} is registered #{count} times. An EventHandler can only be registered once"
+        end
+      end
+    end
+
+    private
+
+    def for_each_autoregisterable_descenant_of(clazz, &block)
+      clazz
+        .descendants
+        .reject(&:abstract_class)
+        .reject(&:skip_autoregister)
+        .each(&block)
     end
   end
 end